requirements.txt
python-dotenv
numpy
google-generativeai
psycopg2-binary
requests
python-telegram-bot[job-queue]
ddgs
pytz
